Fixed arithmetic overflow

+++ src/cluster/DotNext.Net.Cluster/Net/Cluster/Consensus/Raft/RaftCluster.cs
@@ -559,7 +559,7 @@ public abstract partial class RaftCluster<TMember> : Disposable, IRaftCluster, I
                         votes--;
                         break;
                     case PreVoteResult.RejectedByLeader:
-                        votes = int.MinValue;
+                        votes = ushort.MinValue;
                         break;
                 }
             }
